Refine Spinner config logic
